# cherry-pick: storage: disable key access assertions in non-race builds #16131