Discovered in #16068. Using a SpanSetBatch incurs a `O(#spans)` overhead for
each engine operation, which quickly results in quadratic runtime for Raft
command evaluation.

Benchmark:

```
make bench PKG=./pkg/sql/ BENCHES='BenchmarkSQL/^Cockroach/Insert$$/count=100' TESTFLAGS='-count 10 -cpuprofile cpu.out -benchmem' 2>/dev/null | tee $(git rev-parse HEAD)
```

```
benchstat 61d1ee4a3b7eef3aff4d6bf25649aef15dad806f ad1116d51f2c4928857af45478d12d4511d13615
name                               old time/op    new time/op    delta
SQL/Cockroach/Insert/count=100-4     2.50ms ±14%    2.28ms ±31%   -8.53%  (p=0.013 n=10+9)
SQL/Cockroach/Insert/count=1000-4    33.5ms ± 7%    18.4ms ±17%  -45.19%  (p=0.000 n=10+10)

name                               old alloc/op   new alloc/op   delta
SQL/Cockroach/Insert/count=100-4     1.13MB ± 1%    1.13MB ± 1%     ~     (p=0.190 n=10+10)
SQL/Cockroach/Insert/count=1000-4    10.7MB ± 2%    10.5MB ± 0%   -1.89%  (p=0.000 n=10+10)

name                               old allocs/op  new allocs/op  delta
SQL/Cockroach/Insert/count=100-4      2.91k ±37%     2.91k ±34%     ~     (p=0.670 n=10+10)
SQL/Cockroach/Insert/count=1000-4     30.4k ±34%     20.7k ± 0%  -31.94%  (p=0.000 n=10+7)
```
questing a pull to cockroachdb:release-1.0 from cockroachdb:cp-disable-assertions