@mberhault 2017-05-30
CC: @cockroachdb/release